grammar fixes
The change to in the "split panes" section is arguably not a correction, but I think it makes the sentence clearer and easier to understand.  Since the sentence begins by discussing a hypothetical action that the user *can* perform, to switch to describing it as a current reality ("each one of which *shows*") seemed a little confusing to me.