Improve stty handling in it2check and it2getvar
This copies the more robust handling of stty from it2ul to it2check and
it2getvar. Trapping the stty restoration on EXIT means we're covered in
all situations that might cause the script to exit early, not just the
user hitting Ctrl+C, and means we don't need to remember to do it in the
script's happy path. Accordingly, this fixes an issue in which it2getvar
did not restore the stty settings if execution completed without issue.